+++ DoRunDoRun/Sources/Scene/OverallGoalList/OverallGoalListViewController.swift
@@ -156,399 +156,17 @@ extension OverallGoalListViewController: OverallGoalListDisplayLogic {
     }
 }
 
-// MARK: - GoalHeaderView
-final class GoalHeaderView: UIView {
-    
-    // MARK: UI
-    
-    private let iconImageView: UIImageView = {
-        let imageView = UIImageView()
-        imageView.backgroundColor = .init(hex: 0xD7DBE3)
-        imageView.layer.cornerRadius = 16
-        imageView.clipsToBounds = true
-        return imageView
-    }()
-    
-    private let titleLabel = UILabel()
-    
-    private let optionButton: UIButton = {
-        var config = UIButton.Configuration.plain()
-        let imageConfig = UIImage.SymbolConfiguration(pointSize: 12, weight: .regular)
-        if let image = UIImage(systemName: "ellipsis", withConfiguration: imageConfig) {
-            let rotatedImage = image.withRenderingMode(.alwaysOriginal).rotate(radians: .pi / 2)
-            config.image = rotatedImage
-        }
-        
-        config.imagePadding = 0
-        config.baseForegroundColor = UIColor(hex: 0x585D64)
-        
-        let button = UIButton(configuration: config, primaryAction: UIAction { _ in
-            print("Button Click")
-        })
-        return button
-    }()
-    
-    private let currentLabel = UILabel()
-    
-    private let totalLabel = UILabel()
-    
-    private lazy var progressStack: UIStackView = {
-        let stack = UIStackView(arrangedSubviews: [currentLabel, totalLabel])
-        stack.axis = .horizontal
-        stack.spacing = 4
-        return stack
-    }()
-    
-    private let progressView: UIProgressView = {
-        let progress = UIProgressView(progressViewStyle: .default)
-        progress.trackTintColor = .init(hex: 0xD7DBE3)
-        progress.progressTintColor = .init(hex: 0x3E4FFF)
-        progress.layer.cornerRadius = 5
-        progress.clipsToBounds = true
-        return progress
-    }()
-    
-    // MARK: Object lifecycle
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setupView()
-    }
-    
-    @available(*, unavailable)
-    required init?(coder: NSCoder) { fatalError("init(coder:) has not been implemented") }
-    
-    // MARK: Setup
-    
-    private func setupView() {
-        backgroundColor = .init(hex: 0xEDF2FF)
-        layer.cornerRadius = 16
-        
-        [iconImageView, titleLabel, optionButton, progressStack, progressView].forEach {
-            addSubview($0)
-            $0.translatesAutoresizingMaskIntoConstraints = false
-        }
-        
-        NSLayoutConstraint.activate([
-            iconImageView.topAnchor.constraint(equalTo: topAnchor, constant: 20),
-            iconImageView.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 20),
-            iconImageView.widthAnchor.constraint(equalToConstant: 32),
-            iconImageView.heightAnchor.constraint(equalToConstant: 32),
-            
-            titleLabel.centerYAnchor.constraint(equalTo: iconImageView.centerYAnchor),
-            titleLabel.leadingAnchor.constraint(equalTo: iconImageView.trailingAnchor, constant: 8),
-            
-            optionButton.centerYAnchor.constraint(equalTo: titleLabel.centerYAnchor),
-            optionButton.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -20),
-            optionButton.widthAnchor.constraint(equalToConstant: 24),
-            optionButton.heightAnchor.constraint(equalToConstant: 24),
-            
-            progressStack.topAnchor.constraint(equalTo: iconImageView.bottomAnchor, constant: 16),
-            progressStack.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 20),
-            
-            progressView.topAnchor.constraint(equalTo: progressStack.bottomAnchor, constant: 6),
-            progressView.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 20),
-            progressView.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -20),
-            progressView.heightAnchor.constraint(equalToConstant: 10),
-            
-            progressView.bottomAnchor.constraint(equalTo: bottomAnchor, constant: -20)
-        ])
-    }
-    
-    // MARK: Configure
-    
-    func configure(with viewModel: OverallGoalList.GetOverallGoal.ViewModel.DisplayedOverallGoal) {
-        iconImageView.image = UIImage(systemName: viewModel.iconName)
-        
-        titleLabel.attributedText = .withLetterSpacing(
-            text: viewModel.title,
-            font: .pretendard(size: 18, weight: .bold),
-            px: -0.2,
-            color: .init(hex: 0x232529)
-        )
-        
-        currentLabel.attributedText = .withLetterSpacing(
-            text: viewModel.currentSession,
-            font: .pretendard(size: 16, weight: .bold),
-            px: -0.2,
-            color: .init(hex: 0x3E4FFF)
-        )
-        
-        totalLabel.attributedText = .withLetterSpacing(
-            text: viewModel.totalSession,
-            font: .pretendard(size: 12, weight: .regular),
-            px: -0.2,
-            color: .init(hex: 0x82878F)
-        )
-        
-        progressView.setProgress(viewModel.progress, animated: false)
-    }
-}
 
-// MARK: - GoalSessionCell
-final class GoalSessionCell: UITableViewCell {
-    static let identifier = String(describing: GoalSessionCell.self)
-    
-    // MARK: UI
-    
-    private let containerView: UIView = {
-        let view = UIView()
-        view.backgroundColor = UIColor(hex: 0xFFFFFF)
-        view.layer.cornerRadius = 16
-        view.layer.borderColor = UIColor(hex: 0xDFE4EC).cgColor
-        view.layer.borderWidth = 1
-        view.clipsToBounds = true
-        return view
-    }()
-    
-    private let roundLabel = UILabel()
-    
-    private let distanceLabel = UILabel()
-    
-    private let timeLabel = UILabel()
-    
-    private let paceLabel = UILabel()
-    
-    private let checkButton: UIButton = {
-        let button = UIButton(configuration: .checkmark(isChecked: false))
-        return button
-    }()
-    
-    private let firstDivider: UIView = {
-        let view = UIView()
-        view.backgroundColor = UIColor(hex: 0xDFE4EC)
-        return view
-    }()
-    
-    private let secondDivider: UIView = {
-        let view = UIView()
-        view.backgroundColor = UIColor(hex: 0xDFE4EC)
-        return view
-    }()
-    
-    private let retryButton: UIButton = {
-        var config = UIButton.Configuration.filled()
-        let attributedTitle = NSAttributedString.withLetterSpacing(
-            text: "이 목표로 다시 러닝",
-            font: .pretendard(size: 16, weight: .bold),
-            px: -0.2,
-            color: UIColor(hex: 0xFFFFFF)
-        )
-        
-        config.attributedTitle = AttributedString(attributedTitle)
-        config.background.cornerRadius = 12
-        config.baseBackgroundColor = UIColor(hex: 0x3E4FFF)
-        
-        let button = UIButton(configuration: config)
-        return button
-    }()
-    
-    // MARK: Properties
-    
-    // 텍스트 그룹 레이아웃 가이드
-    private let textGroupGuide = UILayoutGuide()
-    
-    // Retry 버튼 상단/높이 여백 제약
-    private var retryButtonTopConstraint: NSLayoutConstraint!
-    private var retryButtonHeightConstraint: NSLayoutConstraint!
-
-    
-    // MARK: Object lifecyle
-    
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        setupView()
-    }
-    
-    @available(*, unavailable)
-    required init?(coder: NSCoder) { fatalError("init(coder:) has not been implemented") }
-    
-    // MARK: Setup
-    
-    private func setupView() {
-        backgroundColor = .clear
-        contentView.backgroundColor = .clear
-        
-        contentView.addSubview(containerView)
-        containerView.translatesAutoresizingMaskIntoConstraints = false
-        
-        NSLayoutConstraint.activate([
-            containerView.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 6),
-            containerView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 20),
-            containerView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -20),
-            containerView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -6)
-        ])
-        
-        [roundLabel, distanceLabel, firstDivider, timeLabel, secondDivider, paceLabel, checkButton, retryButton].forEach {
-            containerView.addSubview($0)
-            $0.translatesAutoresizingMaskIntoConstraints = false
-        }
-        
-        // 텍스트 그룹 레이아웃 가이드 추가
-        containerView.addLayoutGuide(textGroupGuide)
-        
-        // Retry 버튼 상단/높이 여백 제약 설정
-        retryButtonTopConstraint = retryButton.topAnchor.constraint(greaterThanOrEqualTo: paceLabel.bottomAnchor, constant: 0)
-        retryButtonHeightConstraint = retryButton.heightAnchor.constraint(equalToConstant: 0)
-        
-        NSLayoutConstraint.activate([
-            // 회차 라벨
-            roundLabel.topAnchor.constraint(equalTo: containerView.topAnchor, constant: 16),
-            roundLabel.leadingAnchor.constraint(equalTo: containerView.leadingAnchor, constant: 16),
-            
-            // 거리
-            distanceLabel.topAnchor.constraint(equalTo: roundLabel.bottomAnchor, constant: 4),
-            distanceLabel.leadingAnchor.constraint(equalTo: roundLabel.leadingAnchor),
-            
-            // firstDivider
-            firstDivider.centerYAnchor.constraint(equalTo: distanceLabel.centerYAnchor),
-            firstDivider.leadingAnchor.constraint(equalTo: distanceLabel.trailingAnchor, constant: 8),
-            firstDivider.widthAnchor.constraint(equalToConstant: 1),
-            firstDivider.heightAnchor.constraint(equalToConstant: 14),
-            
-            // 시간
-            timeLabel.centerYAnchor.constraint(equalTo: distanceLabel.centerYAnchor),
-            timeLabel.leadingAnchor.constraint(equalTo: firstDivider.trailingAnchor, constant: 8),
-            
-            // secondDivider
-            secondDivider.centerYAnchor.constraint(equalTo: timeLabel.centerYAnchor),
-            secondDivider.leadingAnchor.constraint(equalTo: timeLabel.trailingAnchor, constant: 8),
-            secondDivider.widthAnchor.constraint(equalToConstant: 1),
-            secondDivider.heightAnchor.constraint(equalToConstant: 14),
-            
-            // 페이스
-            paceLabel.centerYAnchor.constraint(equalTo: timeLabel.centerYAnchor),
-            paceLabel.leadingAnchor.constraint(equalTo: secondDivider.trailingAnchor, constant: 8),
-            
-            // 텍스트 그룹 레이아웃 가이드
-            textGroupGuide.topAnchor.constraint(equalTo: roundLabel.topAnchor),
-            textGroupGuide.bottomAnchor.constraint(equalTo: paceLabel.bottomAnchor),
-            
-            // 체크 버튼 (텍스트 그룹 중앙에 정렬)
-            checkButton.centerYAnchor.constraint(equalTo: textGroupGuide.centerYAnchor),
-            checkButton.trailingAnchor.constraint(equalTo: containerView.trailingAnchor, constant: -16),
-            checkButton.widthAnchor.constraint(equalToConstant: 28),
-            checkButton.heightAnchor.constraint(equalToConstant: 28),
-            
-            // 재시도 버튼
-            retryButtonTopConstraint,
-            retryButton.leadingAnchor.constraint(equalTo: containerView.leadingAnchor, constant: 16),
-            retryButton.trailingAnchor.constraint(equalTo: containerView.trailingAnchor, constant: -16),
-            retryButtonHeightConstraint,
-            retryButton.bottomAnchor.constraint(equalTo: containerView.bottomAnchor, constant: -16)
-        ])
-    }
-    
-    // MARK: Configure
-    
-    func configure(with viewModel: OverallGoalList.LoadSessionGoals.ViewModel.DisplayedSessionGoal) {
-        roundLabel.attributedText = .withLetterSpacing(
-            text: viewModel.round,
-            font: .pretendard(size: 16, weight: .bold),
-            px: -0.2,
-            color: .init(hex: 0x232529)
-        )
-        
-        distanceLabel.attributedText = .withLetterSpacing(
-            text: viewModel.distance,
-            font: .pretendard(size: 14, weight: .regular),
-            px: -0.2,
-            color: .init(hex: 0x82878F)
-        )
-        
-        timeLabel.attributedText = .withLetterSpacing(
-            text: viewModel.time,
-            font: .pretendard(size: 14, weight: .regular),
-            px: -0.2,
-            color: .init(hex: 0x232529)
-        )
-        
-        paceLabel.attributedText = .withLetterSpacing(
-            text: viewModel.pace,
-            font: .pretendard(size: 14, weight: .regular),
-            px: -0.2,
-            color: .init(hex: 0x232529)
-        )
-        
-        checkButton.configuration = .checkmark(isChecked: viewModel.isCompleted)
-    }
-    
-    func showRetryButton(_ isVisible: Bool) {
-        retryButton.alpha = isVisible ? 1 : 0
-        retryButtonTopConstraint.constant = isVisible ? 20 : 0
-        retryButtonHeightConstraint.constant = isVisible ? 56 : 0
-        setNeedsLayout()
-    }
-}
-
-final class ToastView: UIView {
-    private let iconView = UIImageView()
-    private let messageLabel = UILabel()
-    
-    init(message: String, style: ToastStyle) {
-        super.init(frame: .zero)
-        setupView(message: message, style: style)
-    }
-    
-    @available(*, unavailable)
-    required init?(coder: NSCoder) { fatalError("init(coder:) has not been implemented") }
-    
-    private func setupView(message: String, style: ToastStyle) {
-        backgroundColor = style.backgroundColor
-        layer.cornerRadius = 12
-        clipsToBounds = true
-        
-        iconView.image = style.icon
-        iconView.tintColor = style.tintColor
-        iconView.contentMode = .scaleAspectFit
-        
-        messageLabel.attributedText = .withLetterSpacing(
-            text: message,
-            font: .pretendard(size: 16, weight: .medium),
-            px: -0.2,
-            color: .init(hex: 0xFFFFFF)
-        )
-        
-        [iconView, messageLabel].forEach {
-            addSubview($0)
-            $0.translatesAutoresizingMaskIntoConstraints = false
-        }
-        
-        NSLayoutConstraint.activate([
-            iconView.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 20),
-            iconView.centerYAnchor.constraint(equalTo: centerYAnchor),
-            iconView.widthAnchor.constraint(equalToConstant: 20),
-            iconView.heightAnchor.constraint(equalToConstant: 20),
-            
-            messageLabel.leadingAnchor.constraint(equalTo: iconView.trailingAnchor, constant: 8),
-            messageLabel.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -20),
-            messageLabel.topAnchor.constraint(equalTo: topAnchor, constant: 12),
-            messageLabel.bottomAnchor.constraint(equalTo: bottomAnchor, constant: -12)
-        ])
-    }
-}
-
-enum ToastStyle {
-    case error
-    
-    var icon: UIImage? {
-        switch self {
-        case .error: return UIImage(systemName: "exclamationmark.circle.fill")
-        }
-    }
-    
-    var tintColor: UIColor {
-        switch self {
-        case .error: return .init(hex: 0xFF443B)
-        }
-    }
-    
-    var backgroundColor: UIColor {
-        switch self {
-        case .error: return .init(hex: 0x232529, alpha: 0.6)
-        }
-    }
-}
+// MARK: - Extensions (임시)
+//
+// 아래 익스텐션들은 원래 별도의 파일로 분리되어야 함.
+// 충돌 방지를 위해 현재는 ViewController 파일에 임시로 두었음.
+// 이후 정리 시에는 각각의 Extension 전용 파일로 이동 필요:
+//
+// - UIImage+.swift
+// - UIButton.Configuration+.swift
+// - UIViewController+.swift
+//
 
 extension UIImage {
     func rotate(radians: CGFloat) -> UIImage {
@@ -603,7 +221,7 @@ extension UIButton.Configuration {
 
 extension UIViewController {
     func showToast(message: String, style: ToastStyle = .error, delay: TimeInterval = 1.0) {
-        let toast = ToastView(message: message, style: style)
+        let toast = InfoToastView(message: message, style: style)
         toast.alpha = 0
         
         view.addSubview(toast)
